make the imageBackground scroll
