+++ views/Search.jsx
@@ -1,12 +1,14 @@
-import { Text, SafeAreaView, ScrollView } from 'react-native';  
+import { Text, SafeAreaView, ScrollView, ImageBackground } from 'react-native';  
 import Header from '../components/Header';
 
-const Search = ({navigation}) => {
+const Search = ({navigation, background}) => {
     return (
       <SafeAreaView>
         <ScrollView contentInsetAdjustmentBehavior="automatic">
-          <Header actualRoute="Search" navigation={navigation} />
-          <Text>Search</Text>
+          <ImageBackground source={background} style={{minHeight: 650}}>
+            <Header actualRoute="Search" navigation={navigation} />
+            <Text>Search</Text>
+          </ImageBackground>
         </ScrollView>
       </SafeAreaView>
     );
